+++ src/test/resources/fixtures/sorting/svelte/style-apply/expected.svelte
@@ -1,0 +1,5 @@
+<style>
+  .button {
+    @apply flex bg-red-500 p-4 text-center font-bold;
+  }
+</style>