+++ src/test/resources/fixtures/sorting/svelte/style-apply/input.svelte
@@ -1,0 +1,5 @@
+<style>
+  .button {
+    @apply text-center p-4 flex bg-red-500 font-bold;
+  }
+</style>